:raising_hand: move skill stuff into a subapp of user
