+++ api/hackathonmentors/user/models.py
@@ -11,13 +11,3 @@ class CustomUser(AbstractUser):
 
     class Meta:
         db_table = 'users'
-
-
-class Skill(models.Model):
-    name = models.CharField(max_length=32)
-
-
-class UserSkill(models.Model):
-    user = models.ForeignKey(CustomUser, on_delete=models.CASCADE)
-    skill = models.ForeignKey(Skill, on_delete=models.CASCADE)
-    competency = models.PositiveSmallIntegerField()
