+++ api/hackathonmentors/user/skill/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class SkillConfig(AppConfig):
+    name = 'skill'
